PanelNum and Patch patterns for testing and emergency fixes.
PanelNum allows for specifying a panel num which will be turned white.  Can be used to help with panel remapping.  Patch can be used to knock out some bad columns of leds that might be flickering (use Subtract blend mode on a channel containing just that pattern).   panelmap.lxp should be used for panel remapping.  It contains a channel for ExPanelTest and another channel for PanelNum. Use PanelNum to navigate to a scrambled panel and then swap the first and second inputs for that panel num.  for example, swap the values in 5.0 and 5.1.  It should unscramble the panel.