Ability to swap first and second inputs per panel.  Ability to swap panels.
This change adds a button to swap the first and second inputs on a panel.  This will swap the universe numbers, save the configuration, and restart the ArtNet output when clicking a button.

There is also the ability to select two panels and click swap panels.  This will swap firstPanel.0 with secondPanel.0 and firstPanel.1 with secondPanel.1.