H264 Video support
Videos should be 420x30 pixels.  Videos can be any length as we stream the frames from a file and seek back to the head of file to play in a loop. Technically, the code can handle videos larger than 420x30 but it is slow since it will do frame resizing in Java at runtime.  Also, currently it will just crop from the middle of the frame whereas the videocrop.sh ffmpeg script allows for specifying the Y offset.

videocrop.sh can take a larger video and produce a 420x30 video.  Some special care needs to be taken for mobile portrait orientation videos. See inside script for details.