Remove alpha anti-aliasing from PGTexture rendering.
This was bleeding a little alpha in the computations which ended up making the image slightly darker than it should be.  Not sure why I tried to add this in the first place.  Something about animated sprites for PGTexture but nobody ever uses that.  Might be better to just threshold hold it full on or full off based on some cut-off value if anybody needs to render an anti-aliased PGTexture pattern over the top of another pattern with the 'Normal' channel mixing mode.