UIModeSelector tweak for better text pattern handling.
Normally, the Standard mode selector will continuously reset the current running time of a channel with a text pattern playing to avoid fading out in the middle of the pattern.  A side effect is that a channel of just text patterns can stall the scheduler.  The fix is to intersperse the Flags pattern and have the scheduler set the current playing time to the configured channel time minus 5 seconds.  This has the effect of transitioning the channel once the channel gets to a Flags pattern (assuming the pattern play time is >5 seconds).