SToyMask aka ShaderToy with a mask image.
The mask can be selected with the mask knob.  It looks for files matching imgpp/stoymask#.png.  This is a more UI conservative technique compared to other file selections in other patterns.  Harder to remember, but file selection takes a lot of space and the Options menu doesn't always render upwards even when requesting it to.