Add positioning and scaling options to RainbowImageBase/PP.
This is handy for prototype a pattern that might have multiple image layers that need to be composited.  Combined with AnimatedSpritePP it is possible to compose a complicated scene using a lot of channels.  At which point, you can bake it into a single pattern as I did for HelloLiverpool.