Force enable channel.autoCycleEnabled for the text channel.
This ensures on startup that the MSGTXT channel that has text patterns will have autoCycleEnabled set to true.  It is possible if somebody saves the show file while a text pattern is playing that the file might be saved with autoCycleEnabled set to false for that channel.  Typically a text pattern will re-enable autoCycleEnabled once it has finished playing it's message.  It does this so that the scheduler doesn't change channels while a message is partially displayed since that would be super annoying.