Add adjustable image scroll speed to RainbowImagePP
This will scroll the image from the right until the image is centered on the Rainbow.  Set speed to some negative value to control how fast it moves into position.  Adjust the channel Dissolve time to adjust fade out overlap with incoming text.  The cycle time will control how long the image is displayed before dissolving to the next pattern.