Updated Liverpool messaging shows.
Added a boolean to RainbowImageBase to enable/disable background alpha on the channel.  When rendering to an alpha enabled background, we would get a visual pop when going into the dissolve.  By rendering to a background without alpha support, we don't experience the background pop.  So the National Museums Liverpool wave graphic pattern should have background alpha disabled.

Updated TextFx4.txts to have the appropriate messaging.