Collpase channel strips.
The ` key will collapse the channel strips leaving only the focused channel strip at full width.  ~ will expand the channel strips back to full width.  There isn't a mechanism to hook into the UI events for the channel strips, so to see a different channel in full width, you need to navigate focus to it (click or arrow keys) and then press ` again to make that channel strip full width.