Fader-based enable/disable.  And fix for standard mode initialization.
If no standard mode channels are detected we will automatically enter Fader-based channel enable/disable.  If the fader value falls to some low threshold (0.05) then disable the channel and set the fader to 0.  If the fader is above that threshold, then enable the channel.

The value is set in StandardModeCycle.prevChannelDisableThreshold.

Not sure if there is some weird internal lifecycle issues with loading new project files, but finally got standard mode working when loading new files.