Add EQ bands texture for shadertoy shaders.
EQ bands are passed on iChannel3.  We are already passing full FFT and audio data on iChannel0, but this is similar to ShaderToyAudio and just passes in the 16 eq bands which is useful.  Technically they can be recomputed by aggregating FFT bins, but that would be pretty wasteful.  Changes rgbaudio.frag to use Eq bands texture.